Removed Comments and made attack interface
The refactor for the comments looks good to me but the attack interface is not done yet. The attack interface has been made but the the other classes havent had all the changes yet except i removed the compareto interface from  DungeonCharacter and instead made it implement attack interface. Program still runs the same so i need to figure out how to make the attack mechanic go off the interface instead of go off the inherited attack method from DungeonCharacter.