Attempting proper attack interface
I dont know if this is right at all. The interface is implemented into dungeon character and the attack method extends the interface but I dont know if that actually makes any difference. The code still runs the same and doesnt have any problems but I don't think the interface is doing anything.